Refactor: split up functionality of the hot reload
Also updated the reload to also check for deleted files
