+++ src/main/kotlin/com/hytaledocs/intellij/hotReload/FileChangeClassifier.kt
@@ -1,0 +1,9 @@
+package com.hytaledocs.intellij.hotReload
+
+interface FileChangeClassifier {
+    fun classify(
+        absolutePath: String,
+        projectBasePath: String,
+        isDeleted: Boolean,
+    ): FileChangeType?
+}
