+++ src/main/kotlin/com/hytaledocs/intellij/hotReload/FileChangeType.kt
@@ -1,0 +1,15 @@
+package com.hytaledocs.intellij.hotReload
+
+sealed class FileChangeType {
+
+    data class SyncWrite(
+        val absoluteSourcePath: String,
+        val absoluteTargetPath: String,
+    ) : FileChangeType()
+
+    data class SyncDelete(
+        val absoluteTargetPath: String,
+    ) : FileChangeType()
+
+    data object SourceCodeChanged : FileChangeType()
+}
