+++ src/main/kotlin/com/hytaledocs/intellij/hotReload/FileSynchronizer.kt
@@ -1,0 +1,8 @@
+package com.hytaledocs.intellij.hotReload
+
+import java.io.File
+
+interface FileSynchronizer {
+    fun sync(source: File, target: File)
+    fun delete(target: File)
+}
