+++ src/main/kotlin/com/hytaledocs/intellij/run/HytaleConsole.kt
@@ -1,0 +1,8 @@
+package com.hytaledocs.intellij.run
+
+interface HytaleConsole {
+    fun println(text: String)
+    fun printInfo(text: String)
+    fun printSuccess(text: String)
+    fun printError(text: String)
+}
